The creeps move.
Implemented a basic wandering and target acquisition algorithm on the enemies. No collisions yet.

Refactored the time_delta to switch from ms to s once. Removed the float division from all places
the time delta was used.

Refactored vmath to clean up the truncate and normalize methods, added additional functionality
with the scale method.

In the set up for the game scene, made a new Surface so the enemies could be green instead of white.

Moved the player to layer 1.